+++ client/src/App.tsx
@@ -93,8 +93,7 @@ function App() {
 export default App;
 
 // TODO:
-// add group search feature - EM
-// add feature to leave group - E
+// add leave group modal before clicking leave group - E
 // add feature to remove user from group - E
 // add ai in chat like meta - (sherians was video on it) - M
 // add pdf upoad - M