+++ server/src/modules/group.module.js
@@ -7,8 +7,7 @@ const groupSchema = new mongoose.Schema({
     groupPic: {
       type: String,
       default: "https://imgs.search.brave.com/7yvnOo5p1WkyZWXxyy781yIzQH35ygTpxopFYwd6Tf4/rs:fit:860:0:0:0/g:ce/aHR0cHM6Ly93d3cu/aWNvbnNkYi5jb20v/aWNvbnMvcHJldmll/dy9ncmF5L2dyb3Vw/LXh4bC5wbmc",
-    },
-    messages: [{ type: mongoose.Schema.Types.ObjectId, ref: "GroupMessage" }], 
+    }
   },
   { timestamps: true }
 );